cmd/ducky: iap dump: Add support for memory dumping via CheckCRC
The XMODEM CRC gives us a unique output value for 1-byte inputs, so we
can call CheckCRC() with a length of 1 to read MCU memory 1 byte at
a time.

CheckCRC() can cause the firmware to be erased if it's given the wrong
value, so we try and calculate the correct value based on reverse
engineering of a dump of the IAP code (obtained via this method).
There's no guarantee that the same algorithm is used in all versions of
the keyboard, so reading memory this way is somewhat "risky" in that the
firmware can be erased inadvertently.

Note that the flash pages storing the IAP code are protected, and so
can't be accidentally erased.