cmd/ducky: iap dump: Reset at start to ensure clean-slate
If the device is already in IAP mode, then it's possible we start trying
to dump without a reset. This might have some dangling CRC state left
over from any previous operation, which can cause an unintended
fail/erase.

Force a reset, even if we're already in IAP mode, to avoid this.